changes in support of reporting RTP projects with consistency scores

+++ Database/Tables/stg.scores.sql
@@ -2,7 +2,7 @@ CREATE TABLE [stg].[scores]
 (
 [ID] [bigint] NULL,
 [AppGUID] [varchar] (max) COLLATE SQL_Latin1_General_CP1_CI_AS NULL,
-[MTPID] [float] NULL,
+[MTPID] [bigint] NULL,
 [WebappsID] [varchar] (max) COLLATE SQL_Latin1_General_CP1_CI_AS NULL,
 [Response] [bigint] NULL
 ) ON [PRIMARY]
